refactor:util/http move interfaces and types from index.ts to types.ts
