+++ lib/datasource/docker/index.ts
@@ -8,10 +8,10 @@ import { HOST_DISABLED } from '../../constants/error-messages';
 import { logger } from '../../logger';
 import type { HostRule } from '../../types';
 import { ExternalHostError } from '../../types/errors/external-host-error';
+import type { Http } from '../../util/http';
+import { HttpError, HttpOptions, HttpResponse } from '../../util/http/types';
 import * as packageCache from '../../util/cache/package';
 import * as hostRules from '../../util/host-rules';
-import type { Http, HttpOptions, HttpResponse } from '../../util/http';
-import { HttpError } from '../../util/http/types';
 import type { OutgoingHttpHeaders } from '../../util/http/types';
 import { hasKey } from '../../util/object';
 import { regEx } from '../../util/regex';
