+++ lib/platform/gitlab/index.ts
@@ -21,8 +21,8 @@ import { logger } from '../../logger';
 import { BranchStatus, PrState, VulnerabilityAlert } from '../../types';
 import * as git from '../../util/git';
 import * as hostRules from '../../util/host-rules';
-import type { HttpResponse } from '../../util/http';
 import { setBaseUrl } from '../../util/http/gitlab';
+import type { HttpResponse } from '../../util/http/types';
 import { regEx } from '../../util/regex';
 import { sanitize } from '../../util/sanitize';
 import { ensureTrailingSlash, getQueryString, parseUrl } from '../../util/url';
