+++ lib/util/http/bitbucket-server.ts
@@ -1,6 +1,7 @@
 import { PlatformId } from '../../constants';
 import { resolveBaseUrl } from '../url';
-import { Http, HttpOptions, HttpResponse, InternalHttpOptions } from '.';
+import type { HttpOptions, HttpResponse, InternalHttpOptions } from './types';
+import { Http } from '.';
 
 let baseUrl: string;
 export const setBaseUrl = (url: string): void => {
