+++ lib/util/http/types.ts
@@ -1,3 +1,4 @@
+import type { IncomingHttpHeaders } from 'http';
 import {
   OptionsOfBufferResponseBody,
   OptionsOfJSONResponseBody,
@@ -34,3 +35,40 @@ export interface RequestStats {
 }
 
 export type OutgoingHttpHeaders = Record<string, string | string[] | undefined>;
+
+export interface HttpOptions {
+  body?: any;
+  username?: string;
+  password?: string;
+  baseUrl?: string;
+  headers?: OutgoingHttpHeaders;
+
+  /**
+   * Do not use authentication
+   */
+  noAuth?: boolean;
+
+  throwHttpErrors?: boolean;
+  useCache?: boolean;
+}
+
+export interface HttpPostOptions extends HttpOptions {
+  body: unknown;
+}
+
+export interface InternalHttpOptions extends HttpOptions {
+  json?: Record<string, unknown>;
+  responseType?: 'json' | 'buffer';
+  method?: 'get' | 'post' | 'put' | 'patch' | 'delete' | 'head';
+}
+
+export interface HttpHeaders extends IncomingHttpHeaders {
+  link?: string | undefined;
+}
+
+export interface HttpResponse<T = string> {
+  statusCode: number;
+  body: T;
+  headers: HttpHeaders;
+  authorization?: boolean;
+}
